requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
alembic==1.12.0
asyncpg==0.30.0
databases==0.6.1
pydantic==1.10.11
python-dotenv==1.0.0
httpx==0.24.1
stripe==6.6.0
